feat[database]: update 'users' and update 'articles' ## Changes: + '/models/user.js': created + 'model /models/user.js': created + '/models/articles.js': created + '/seeders/20190521161026-user.js': updated + 'model /models/article.js': created + '/models/articles.js': created + '/models/user.js': created + '/models/article.js': created

+++ models/user.js
@@ -16,7 +16,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false
     },
     status: {
-      type: DataTypes.INTEGER(1),
+      type: DataTypes.INTEGER(11),
       allowNull: false
     },
     createdat: {
